+++ R/rostemplate-package.R
@@ -3,4 +3,5 @@
 
 # import stuffs
 #' @importFrom utils menu
+#' @importFrom grDevices adjustcolor col2rgb colorRampPalette hsv rgb2hsv
 NULL